tp0/extras/requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.5.3
Werkzeug==0.9.6
argparse==1.2.1
itsdangerous==0.24
numpy==1.9.0
wsgiref==0.1.2
